requirements.txt
cycler==0.10.0
decorator==4.4.2
imageio==2.9.0
joblib==0.16.0
kiwisolver==1.2.0
matplotlib==3.3.0
networkx==2.4
numpy~=1.18.5
Pillow==7.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
PyWavelets==1.1.1
scikit-image==0.17.2
scikit-learn==0.23.2
scipy~=1.5.4
six==1.15.0
threadpoolctl==2.1.0
tifffile==2020.7.24

tensorflow-gpu~=2.3.1
sklearn~=0.0
PyYAML~=5.3.1
pandas~=1.1.4